+++ packages/liveblocks-core/src/protocol/AuthToken.ts
@@ -1,8 +1,8 @@
 import { isPlainObject } from "../lib/guards";
 import type { Json } from "../lib/Json";
 import { b64decode, tryParseJson } from "../lib/utils";
-import type { Permission } from "../permissions";
 import type { IUserInfo } from "./BaseUserMeta";
+import type { Permission } from "./Permissions";
 
 export type LiveblocksPermissions = Record<string, Permission[]>;
