+++ packages/liveblocks-core/src/refs/ManagedOthers.ts
@@ -2,7 +2,7 @@ import { freeze } from "../lib/freeze";
 import type { JsonObject } from "../lib/Json";
 import { DerivedSignal, merge, MutableSignal } from "../lib/signals";
 import { compact, compactObject } from "../lib/utils";
-import { hasPermissionCapability } from "../permissions";
+import { hasPermissionCapability } from "../protocol/Permissions";
 import type { BaseUserMeta } from "../protocol/BaseUserMeta";
 import type { User } from "../types/User";
